finish implementation

+++ src/app/controllers/CardController.ts
@@ -1,6 +1,5 @@
 import { Request, Response } from 'express';
 import { getRepository } from 'typeorm';
-import { v4 as uuid } from 'uuid';
 
 import Card from '../models/Card';
 
@@ -16,26 +15,27 @@ class CardController {
         const repository = getRepository(Card);
 
         const newCard = {
-            id: uuid(),
             title: title,
             content: content,
             attend: false,
             created_at: new Date(),
             updated_at: new Date(),
         }
-        const card = repository.create(newCard);
+        const card = await repository.create(newCard);
 
         await repository.save(card);
         return response.json(card);
     }
 
     async edit(request: Request, response: Response) {
-        const { id, title, content } = request.body;
+        const { id } = request.params;
+        const { title, content } = request.body;
         const repository = getRepository(Card);
-        const cardExists = await repository.findOne({ where: id });
+        const cardExists = await repository.findOne(id);
 
         if (!cardExists)
-            throw new Error('Could not find card');
+            return response.status(404).json({ error: 'Could not find card' });
+
         cardExists.title = title;
         cardExists.content = content;
 
@@ -45,11 +45,11 @@ class CardController {
 
     async setToDone(request: Request, response: Response) {
         const repository = getRepository(Card);
-        const { id } = request.body;
-        const cardExists = await repository.findOne({ where: id });
+        const { id } = request.params;
+        const cardExists = await repository.findOne(id);
 
         if (!cardExists)
-            throw new Error('Card not exists');
+            return response.status(404).json({ error: 'Could not find card' });
 
         cardExists.attend = true;
 
@@ -59,12 +59,15 @@ class CardController {
 
     async delete(request: Request, response: Response) {
         const repository = getRepository(Card);
-        const { id } = request.body;
-        
-        const cardExists = await repository.findOne({ where: id });
+        const { id } = request.params;
+
+        const cardExists = await repository.findOne(id);
+
+        if (!cardExists)
+            return response.status(404).json({ error: 'Could not find card' });
 
         await repository.remove(cardExists);
-        return response.sendStatus(203);
+        return response.sendStatus(204);
     }
 }
